Added an `url_verifier` to verify versions against download urls
This commit adds the possibility to give an optional url to verify a
version number (for example, it can be tested if a tarball exists for a
specific version number). Urls must contain a placeholder `{}` for the
complete version number (as a string). You can specify an optional
expression within the parentheses to manipulate the version number
before the substitution. As an example, `{x[1:]}` would remove the first
character of each version string.